made `test_upload_and_download_file` readable
the previous version of the test was harder to read with more functions, this version is much simpler and easier to read and maintain.

Although, there are some functions that are lesser known and some code is written as one line instead of multiple.

However, overall it is much clearer and easier to read/maintain.