ignoring mypy error
mypy says that the `logger: logging.Logger = None` this cannot be None because you said it should be `logging.Logger` but before it is initialized it would be None.

Don't want to type it as `Optional` because it should always exist and it is not Optional, so just ignoring the error for now.